pr comments fix
