+++ projects/core/src/auth/client-auth/store/effects/client-token.effect.ts
@@ -9,7 +9,7 @@ import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { Observable, of } from 'rxjs';
 import { catchError, exhaustMap, map } from 'rxjs/operators';
 import { LoggerService } from '../../../../logger';
-import { tryNormalizeHttpError } from '../../../../util/normalize-http-error';
+import { tryNormalizeHttpError } from '../../../../util/try-normalize-http-error';
 import { ClientToken } from '../../../client-auth/models/client-token.model';
 import { ClientAuthenticationTokenService } from '../../services/client-authentication-token.service';
 import { ClientAuthActions } from '../actions/index';
