+++ projects/core/src/error-handling/effects-error-handler/cx-error-handler.effect.ts
@@ -6,9 +6,8 @@
 
 import { Injectable } from '@angular/core';
 import { Actions, createEffect } from '@ngrx/effects';
-import { filter, map, tap } from 'rxjs/operators';
+import { filter, tap } from 'rxjs/operators';
 import { EffectsErrorHandlerService } from './effects-error-handler.service';
-import { Action } from '@ngrx/store';
 import { ErrorAction } from '@spartacus/core';
 import { Observable } from 'rxjs';
 
@@ -17,13 +16,10 @@ export class CxErrorHandlerEffect {
   error$: Observable<ErrorAction> = createEffect(
     () =>
       this.actions$.pipe(
-        filter((action: Action) =>
-          this.effectErrorHandler.filterActions(action)
-        ),
-        map((action: Action) => action as ErrorAction),
-        tap((action: ErrorAction) => {
-          this.effectErrorHandler.handleError(action);
-        })
+        filter(this.effectErrorHandler.filterActions),
+        tap((errorAction: ErrorAction) =>
+          this.effectErrorHandler.handleError(errorAction)
+        )
       ),
     { dispatch: false }
   );
