+++ projects/core/src/product/store/effects/product.effect.ts
@@ -12,7 +12,7 @@ import { catchError, map, mergeMap } from 'rxjs/operators';
 import { AuthActions } from '../../../auth/user-auth/store/actions';
 import { LoggerService } from '../../../logger';
 import { SiteContextActions } from '../../../site-context/store/actions/index';
-import { tryNormalizeHttpError } from '../../../util/normalize-http-error';
+import { tryNormalizeHttpError } from '../../../util/try-normalize-http-error';
 import { bufferDebounceTime } from '../../../util/rxjs/buffer-debounce-time';
 import { withdrawOn } from '../../../util/rxjs/withdraw-on';
 import { ProductConnector } from '../../connectors/product/product.connector';
