+++ projects/core/src/site-context/store/effects/languages.effect.ts
@@ -16,7 +16,7 @@ import {
   map,
 } from 'rxjs/operators';
 import { LoggerService } from '../../../logger';
-import { tryNormalizeHttpError } from '../../../util/normalize-http-error';
+import { tryNormalizeHttpError } from '../../../util/try-normalize-http-error';
 import { SiteConnector } from '../../connectors/site.connector';
 import { SiteContextActions } from '../actions/index';
 import { getActiveLanguage } from '../selectors/languages.selectors';
