+++ projects/core/src/user/store/effects/delivery-countries.effect.ts
@@ -11,7 +11,7 @@ import { catchError, map, switchMap } from 'rxjs/operators';
 import { LoggerService } from '../../../logger';
 import { CountryType } from '../../../model/address.model';
 import { SiteConnector } from '../../../site-context/connectors/site.connector';
-import { tryNormalizeHttpError } from '../../../util/normalize-http-error';
+import { tryNormalizeHttpError } from '../../../util/try-normalize-http-error';
 import { UserActions } from '../actions/index';
 
 @Injectable()
