+++ projects/core/src/user/store/effects/regions.effect.ts
@@ -12,7 +12,7 @@ import { catchError, map, switchMap } from 'rxjs/operators';
 import { LoggerService } from '../../../logger';
 import { SiteConnector } from '../../../site-context/connectors/site.connector';
 import { StateUtils } from '../../../state/utils/index';
-import { tryNormalizeHttpError } from '../../../util/normalize-http-error';
+import { tryNormalizeHttpError } from '../../../util/try-normalize-http-error';
 import { UserActions } from '../actions/index';
 import { REGIONS } from '../user-state';
