+++ projects/core/src/util/index.ts
@@ -12,6 +12,7 @@ export * from './glob.service';
 export * from './http-params-uri.encoder';
 export * from './java-reg-exp-converter/index';
 export * from './normalize-http-error';
+export * from './try-normalize-http-error';
 export * from './occ-http-error-constants';
 export * from './occ-http-error-handlers';
 export * from './regex-pattern';
