+++ projects/core/src/util/normalize-http-error.ts
@@ -57,10 +57,3 @@ export function normalizeHttpError(
 
   return undefined;
 }
-
-export function tryNormalizeHttpError(
-  error: HttpErrorResponse | HttpErrorModel | any,
-  logger?: LoggerService
-): HttpErrorModel | Error {
-  return normalizeHttpError(error, logger) ?? error;
-}
