+++ projects/core/src/util/try-normalize-http-error.ts
@@ -1,0 +1,13 @@
+import { HttpErrorResponse } from '@angular/common/http';
+import {
+  HttpErrorModel,
+  LoggerService,
+  normalizeHttpError,
+} from '@spartacus/core';
+
+export function tryNormalizeHttpError(
+  error: HttpErrorResponse | HttpErrorModel | any,
+  logger?: LoggerService
+): HttpErrorModel | Error {
+  return normalizeHttpError(error, logger) ?? error;
+}
